Version 1.0.26.1 - the cities probably work, however, I don't guarantee it and I'm not sure.
Now I added that the settlement will be deleted and the option to delete buildings, so when a player leaves the game all its buildings are being deleted. in addition, the settlement will be deleted to being a city, so there is a city and not a settlement. Therefore, I added this option and I don't know if it works perfectly.